% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_integration_response}
\alias{apigatewayv2_get_integration_response}
\title{Gets an IntegrationResponses}
\usage{
apigatewayv2_get_integration_response(ApiId, IntegrationId,
  IntegrationResponseId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}

\item{IntegrationResponseId}{[required] The integration response ID.}
}
\description{
Gets an IntegrationResponses.
}
\section{Request syntax}{
\preformatted{apigatewayv2$get_integration_response(
  ApiId = "string",
  IntegrationId = "string",
  IntegrationResponseId = "string"
)
}
}

\keyword{internal}
