% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_update_api}
\alias{apigatewayv2_update_api}
\title{Updates an Api resource}
\usage{
apigatewayv2_update_api(ApiId, ApiKeySelectionExpression, Description,
  DisableSchemaValidation, Name, RouteSelectionExpression, Version)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ApiKeySelectionExpression}{An API key selection expression. See \href{https://docs.aws.amazon.com/apigateway/latest/developerguide/apigateway-websocket-api-selection-expressions.html#apigateway-websocket-api-apikey-selection-expressions}{API Key Selection Expressions}.}

\item{Description}{The description of the API.}

\item{DisableSchemaValidation}{Avoid validating models when creating a deployment.}

\item{Name}{The name of the API.}

\item{RouteSelectionExpression}{The route selection expression for the API.}

\item{Version}{A version identifier for the API.}
}
\description{
Updates an Api resource.
}
\section{Request syntax}{
\preformatted{apigatewayv2$update_api(
  ApiId = "string",
  ApiKeySelectionExpression = "string",
  Description = "string",
  DisableSchemaValidation = TRUE|FALSE,
  Name = "string",
  RouteSelectionExpression = "string",
  Version = "string"
)
}
}

\keyword{internal}
